# Apply `Array.map` and `Array.forEach` alts