--- a/package.json
+++ b/package.json
@@ -29,6 +29,7 @@
     "format": "prettier-eslint --write \"src/**/*.js\" \"test/**/*.js\" \"example/**/*.js\" \"jest-utils/**/*.js\""
   },
   "dependencies": {
+    "@arr/map": "^0.1.0",
     "babel-plugin-syntax-jsx": "^6.18.0",
     "styled-components": "2.0.0",
     "theming": "^1.0.1",