--- a/src/index.js
+++ b/src/index.js
@@ -1,5 +1,6 @@
 // @flow
 import map from '@arr/map'
+import forEach from '@arr/foreach'
 import { StyleSheet } from './sheet'
 import { hashArray, hashObject } from './hash'
 
@@ -47,8 +48,10 @@ export function css (classes: string[], vars: vars, content: () => string[]) {
 
     if (!inserted[hash]) {
       inserted[hash] = true
-      map(src, r => r.replace(new RegExp(classes[0], 'gm'), `${classes[0]}-${hash}`))
-        .forEach(r => sheet.insert(r))
+      forEach(
+        map(src, r => r.replace(new RegExp(classes[0], 'gm'), `${classes[0]}-${hash}`)),
+        r => sheet.insert(r)
+      )
     }
     return `${classes[0]}-${hash} ${computedClassName}`
   }
@@ -63,7 +66,7 @@ export function injectGlobal (src: string[]) {
   const hash = hashArray(src)
   if (!inserted[hash]) {
     inserted[hash] = true
-    src.forEach(r => sheet.insert(r))
+    forEach(src, r => sheet.insert(r))
   }
 }
 
@@ -74,7 +77,7 @@ export function keyframes (kfm: string, src: string[]) {
   const animationName = `${kfm}-${hash}`
   if (!inserted[hash]) {
     inserted[hash] = true
-    src.forEach(r => {
+    forEach(src, r => {
       sheet.insert(`@keyframes ${animationName} ${r}`)
     })
   }
@@ -82,7 +85,7 @@ export function keyframes (kfm: string, src: string[]) {
 }
 
 export function hydrate (ids: string[]) {
-  ids.forEach(id => (inserted[id] = true))
+  forEach(ids, id => (inserted[id] = true))
 }
 
 // 🍩
@@ -95,7 +98,7 @@ export function objStyle (style: { [string]: any }) {
   if (inserted[hash]) return className
 
   const rules = deconstruct(selector, style)
-  rules.forEach(rule => sheet.insert(rule))
+  forEach(rules, rule => sheet.insert(rule))
 
   inserted[hash] = true
 
@@ -114,18 +117,18 @@ function deconstruct (selector, styles, media) {
       decs.push(createDec(key, value))
       continue
     } else if (Array.isArray(value)) {
-      value.forEach(val => {
+      forEach(value, val => {
         decs.push(createDec(key, val))
       })
       continue
     } else if (/^:/.test(key)) {
-      deconstruct(selector + key, value, media).forEach(r => rules.push(r))
+      forEach(deconstruct(selector + key, value, media), r => rules.push(r))
       continue
     } else if (/^@media/.test(key)) {
-      deconstruct(selector, value, key).forEach(r => rules.push(r))
+      forEach(deconstruct(selector, value, key), r => rules.push(r))
       continue
     } else {
-      deconstruct(selector + ' ' + key, value, media).forEach(r =>
+      forEach(deconstruct(selector + ' ' + key, value, media), r =>
         rules.push(r)
       )
       continue